# LZ4 decompression forward compatibility